% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buy_me_stuff.R
\name{buy_me_stuff_btn}
\alias{buy_me_stuff_btn}
\title{Randomly select if donate button will be alcoholic}
\usage{
buy_me_stuff_btn()
}
\value{
name of HTML file to use for button
}
\description{
Randomly select if donate button will be alcoholic
}
\keyword{internal}
