% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_list.R
\name{get_match_list}
\alias{get_match_list}
\title{Return list of pattern matches and captured text}
\usage{
get_match_list(
  str,
  pattern,
  ignore_case = TRUE,
  global = TRUE,
  perl = TRUE,
  fixed = FALSE
)
}
\arguments{
\item{str}{string to apply regex to}

\item{pattern}{regex to search for in str}

\item{ignore_case}{see ?gsub}

\item{global}{see ?gsub}

\item{perl}{see ?gsub}

\item{fixed}{see ?gsub}
}
\value{
a list with names as matched text and elements as
        character vectors of matched text
}
\description{
Return list of pattern matches and captured text
}
\examples{
# get_match_list('abc aaa abc acdc', 'a(.)(.)')

# list(abc = c("b", "c"),
#      aaa = c("a", "a"),
#      abc = c("b", "c"),
#      acd = c("c", "d"))
}
\keyword{internal}
