% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{half_slashes}
\alias{half_slashes}
\title{Half slashes for regular expression matching}
\usage{
half_slashes(str)
}
\arguments{
\item{str}{regular expression string}
}
\description{
Half slashes for regular expression matching
}
