% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{highlight_test_str}
\alias{highlight_test_str}
\title{Function to highlight test string}
\usage{
highlight_test_str(str, pattern, ignore_case = TRUE, global = TRUE,
  perl = TRUE, fixed = FALSE, color_palette = "Set2")
}
\arguments{
\item{str}{Regular expression string}

\item{pattern}{Pattern to be matched}

\item{ignore_case}{Whether or not to ignore cases}

\item{global}{Whether the matches are global expression}

\item{perl}{Whether to use Perl or POSIX standard}

\item{fixed}{...}

\item{color_palette}{Color pallete to be used for highlighing test string}
}
\description{
Function to highlight test string
}
