% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_format_match_list.R
\name{html_format_match_list}
\alias{html_format_match_list}
\title{Create HTML component to show list of items matched by pattern in test str}
\usage{
html_format_match_list(match_list, color_palette = "Set3")
}
\arguments{
\item{match_list}{output of get_match_list()}

\item{color_palette}{RColorBrewer palette name for highlighting colors}
}
\value{
HTML string to be rendered with shiny::HTML()
}
\description{
Create HTML component to show list of items matched by pattern in test str
}
\keyword{internal}
