% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{mgsub}
\alias{mgsub}
\title{Function to substitute test string}
\usage{
mgsub(pattern, replacement, text.var, leadspace = FALSE,
  trailspace = FALSE, fixed = TRUE, trim = TRUE,
  order.pattern = fixed, ...)
}
\arguments{
\item{pattern}{Pattern to be matched}

\item{replacement}{Replacement}

\item{text.var}{...}

\item{leadspace}{...}

\item{trailspace}{...}

\item{fixed}{...}

\item{trim}{...}

\item{order.pattern}{...}

\item{...}{...}
}
\description{
Function to substitute test string
}
