% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexplainer.R
\name{regexplain}
\alias{regexplain}
\title{Function to explain the regex}
\usage{
regexplain(regx)
}
\arguments{
\item{regx}{Regular expression statement in strings}
}
\value{
a dataframe that returns the strings
}
\description{
Function to explain the regex
}
