% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexplain.R
\name{split_vec}
\alias{split_vec}
\title{Spit at a specified list of indices}
\usage{
split_vec(vec, splits)
}
\arguments{
\item{vec}{vector to split}

\item{splits}{vector of split locations}
}
\value{
list of vectors
}
\description{
Spit at a specified list of indices
}
\examples{
# split_vec(1:4, 3)
# list(`0` = 1:2, `1` = 3:4)
}
\keyword{internal}
