% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_default.R
\name{try_default}
\alias{try_default}
\title{Create a 'safe' version of a function with a default return value}
\usage{
try_default(fun, default = NULL, silent = FALSE)
}
\arguments{
\item{fun}{function to add try logic to}

\item{default}{default value to return if fun errors}

\item{silent}{should error be printed to console?}
}
\value{
either output of fun or default value in case of error
}
\description{
Create a 'safe' version of a function with a default return value
}
\examples{
# add = function(a, b) a + b
# default_add = try_default(add)
# null_output = default_add(1, 'a')
}
\keyword{internal}
