% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_match_list}
\alias{get_match_list}
\title{Function to match text using}
\usage{
get_match_list(str, pattern, ignore_case = TRUE, global = TRUE,
  perl = TRUE, fixed = FALSE)
}
\arguments{
\item{str}{Regular expression string}

\item{pattern}{Pattern to be matched}

\item{ignore_case}{Whether or not to ignore cases}

\item{global}{Whether the matches are global expression}

\item{perl}{Whether to use Perl or POSIX standard}

\item{fixed}{...}
}
\description{
Function to match text using
}
