% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/half_slashes.R
\name{half_slashes}
\alias{half_slashes}
\title{Given a string, remove half of the escaping backslashes}
\usage{
half_slashes(str, exclude = character(0))
}
\arguments{
\item{str}{A string to remove backslashes from}

\item{exclude}{A character vector of items to exclude from processing}
}
\value{
A version of the input string with half of the escaping chars
}
\description{
Given a string, remove half of the escaping backslashes
}
\details{
An exception will be thrown if removing slashes leads to a parsing
        error.  For example, "\\" will be converted to "\" which is not
        a valid string.
}
\examples{
# half_slashes("\\\\")
# [1] "\\"
}
\keyword{internal}
