% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_test_str.R
\name{highlight_test_str}
\alias{highlight_test_str}
\title{Highlight regex matches in output display with HTML}
\usage{
highlight_test_str(
  str,
  pattern,
  ignore_case = TRUE,
  global = TRUE,
  perl = TRUE,
  fixed = FALSE,
  color_palette = "Set3"
)
}
\arguments{
\item{str}{string to find matches in}

\item{pattern}{pattern to match in str}

\item{ignore_case}{see ?gsub}

\item{global}{see ?gsub}

\item{perl}{see ?gsub}

\item{fixed}{see ?gsub}

\item{color_palette}{RColorBrewer palette name for highlighting colors}
}
\value{
HTML string to be rendered with shiny::HTML()
}
\description{
Highlight regex matches in output display with HTML
}
\keyword{internal}
