% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{html_format_match_list}
\alias{html_format_match_list}
\title{Function to format matched list}
\usage{
html_format_match_list(match_list, color_palette = "Set2")
}
\arguments{
\item{match_list}{Matched list}

\item{color_palette}{Color pallete to be used for displaying matches}
}
\description{
Function to format matched list
}
