% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgsub.R
\name{mgsub}
\alias{mgsub}
\title{Multi gsub}
\usage{
mgsub(pattern, replacement, text.var, fixed = TRUE, ...)
}
\arguments{
\item{pattern}{Character string to be matched in the given character
vector.}

\item{replacement}{Character string equal in length to pattern or of length
one which are a replacement for matched pattern.}

\item{text.var}{The text variable.}

\item{fixed}{logical. If TRUE, pattern is a string to be matched as is.
Overrides all conflicting arguments.}

\item{...}{Additional arguments passed to gsub.}
}
\value{
text.var with substitutions applied
}
\description{
I wanted qdap::mgsub() w/o having to have full qdap package.
             qdap has had issues related to deploying a shinyapp.  So this
             is a trimmed down version of qdap::mgsub()
}
\keyword{internal}
