% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexplain.R
\name{regexplain}
\alias{regexplain}
\title{Create an 'explanation' of an inputted regular expression}
\usage{
regexplain(regx)
}
\arguments{
\item{regx}{string containing regular expression to explain}
}
\value{
a dataframe with columns c("regex", "explanation")
}
\description{
Create an 'explanation' of an inputted regular expression
}
\details{
Provided by rick.measham.id.au
}
\keyword{internal}
