% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexplainer.R
\name{split_vec}
\alias{split_vec}
\title{Function to}
\usage{
split_vec(vec, splits)
}
\arguments{
\item{vec}{vector of regular expression to be split}

\item{splits}{The splits}
}
\description{
Function to
}
